requirements.txt
## The following requirements were added by pip freeze:
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.19
cffi==1.6.0
contextlib2==0.5.3
cryptography==1.3.2
decorator==4.0.9
dj-database-url==0.3.0
Django==1.8.9
django-celery==3.1.17
django-extensions==1.5.5
django-redis==4.4.3
enum34==1.1.6
gnureadline==6.3.3
gunicorn==19.3.0
hotqueue==0.2.7
idna==2.1
ipaddress==1.0.16
ipython==4.2.0
ipython-genutils==0.1.0
kombu==3.0.35
-e git+https://github.com/geeknam/messengerbot.git@1786695ff43034c3d548e5fd222661480dd4bbb7#egg=messengerbot
ndg-httpsclient==0.4.0
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python-redis-lock==3.0.0
pytz==2016.4
raven==5.18.0
redis==2.10.5
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
